+++ Sources/Nuke/Internal/Formatter.swift
@@ -1,0 +1,17 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2026 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+enum Formatter {
+    static func bytes(_ count: Int) -> String {
+        bytes(Int64(count))
+    }
+
+    static func bytes(_ count: Int64) -> String {
+        let formatter = ByteCountFormatter()
+        formatter.allowsNonnumericFormatting = false // "0 bytes", not "Zero KB"
+        return formatter.string(fromByteCount: count)
+    }
+}